+++ app/controllers/vendalists_controller.rb
@@ -6,7 +6,7 @@ class VendalistsController < ApplicationController
   def vender
     user_id = params[:user_id]
     produto_id = params[:produto_id]
-    quantity = params[:quantity].to_i
+    quantity = params[:quantity].to_f
 
     inventory_list = Inventorylist.find_by(user_id:, produto_id:)
