+++ app/controllers/vendas_controller.rb
@@ -8,7 +8,7 @@ class VendasController < ApplicationController
   def adicionar_uma_venda
     user_id = params[:user_id]
     produto_id = params[:produto_id]
-    quantity = params[:quantity].to_i
+    quantity = params[:quantity].to_f
 
     # aqui ele tem que aceitar varios produtos
     venda = Venda.new(user_id:, produto_id:, quantity:)
